requirements.txt
PyQt5
PyYAML
requests
